+++ modules/client/test/integration/client/encoding.test.ts
@@ -15,7 +15,6 @@ import {
   DaoMetadata,
   IGrantPermissionParams,
   IRevokePermissionParams,
-  LIVE_CONTRACTS,
   Permissions,
   WithdrawParams,
 } from "../../../src";